Legion_Omega_V0.4_CREWAI/requirements.txt
crewai>=1.14,<2.0
crewai-tools>=0.12
litellm>=1.50
python-telegram-bot>=21,<23
pydantic>=2.7
pydantic-settings>=2.2
python-dotenv>=1.0
aiofiles>=23
Pillow>=10
